+++ packages/blocks-core/CHANGELOG.md
@@ -43,6 +43,7 @@
 - `Quote` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `Pullquote` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `Table` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
+- `Term Description` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 
 
 ### Automated Tests
@@ -87,6 +88,7 @@
 - Added E2E tests to check functionality `Quote` new inner block.
 - Added E2E tests to check functionality `Pullquote` new inner block.
 - Added E2E tests to check functionality `Table` new inner block.
+- Added E2E tests to check functionality `Term Description` new inner block.
 
 ## 1.4.0 (2025-05-19)